Added logic to subscribe agency, category and news.
Users will be able to subscribe to any categories and agencies.
Added scope to all the models for retrieving records specific to any
user.
Modified existing services of those model for retrieving both global and
user version of records.
AuthService now has a function which gives the status of logged in.
The global fetch results of categories and agencies in services are
cached for later diff evaluation.